+++ crates/pathfinder/src/bin/pathfinder/main.rs
@@ -243,6 +243,9 @@ Hint: This is usually caused by exceeding the file descriptor limit of your syst
         native_execution: config.native_execution.is_enabled(),
         native_class_cache_size: config.native_execution.class_cache_size(),
         native_compiler_optimization_level: config.native_execution.optimization_level(),
+        native_execution_force_use_for_incompatible_classes: config
+            .native_execution
+            .force_use_for_incompatible_classes(),
         submission_tracker_time_limit: config.submission_tracker_time_limit,
         submission_tracker_size_limit: config.submission_tracker_size_limit,
         block_trace_cache_size: config.rpc_block_trace_cache_size,