+++ crates/rpc/src/method/estimate_fee.rs
@@ -114,6 +114,9 @@ pub async fn estimate_fee(
             context.contract_addresses.eth_l2_token_address,
             context.contract_addresses.strk_l2_token_address,
             context.native_class_cache,
+            context
+                .config
+                .native_execution_force_use_for_incompatible_classes,
         );
 
         let skip_validate = input