+++ crates/rpc/src/method/estimate_message_fee.rs
@@ -118,6 +118,9 @@ pub async fn estimate_message_fee(
             context.contract_addresses.eth_l2_token_address,
             context.contract_addresses.strk_l2_token_address,
             context.native_class_cache,
+            context
+                .config
+                .native_execution_force_use_for_incompatible_classes,
         );
 
         let transaction = create_executor_transaction(input, context.chain_id)?;