+++ crates/rpc/src/method/trace_block_transactions.rs
@@ -145,6 +145,9 @@ pub async fn trace_block_transactions(
             context.contract_addresses.eth_l2_token_address,
             context.contract_addresses.strk_l2_token_address,
             context.native_class_cache,
+            context
+                .config
+                .native_execution_force_use_for_incompatible_classes,
         );
         let traces =
             match pathfinder_executor::trace(db_tx, state, cache, hash, executor_transactions) {